Add files via upload
`SIREN.py` is the helper file containing the SIREN and SIRENLayer classes. `h_SIREN_training.py` imports data from the SXS database (https://data.black-holes.org/simulations/index.html) to be used for training. It currently extracts only the real part of the dominant (2, 2) mode and trims the strain from the built-in `reference_time` to the time of maximum strain (calculated by a norm of all the modes, corresponds pretty well with merger).