Add new files, update preprocessing + training
Updated `h_SIREN_training.py` to remove unnecessary parameters and use AdamW with decaying learning rate (might not be necessary, probably worth messing around with more). The big change is that all the data now starts at a peak of the waveform - without this the simulations start at settle-down times at random points in the wave. The model couldn't converge since the phase didn't depend on any parameters.
Added a new network type (named it SHAN, might have another name) that's basically a MLP until the end. Generates a penultimate layer (A, B, C) that's used to predict the waveform $A \sin(Bt + C)$. That's now the only place $\sin$ is used - the rest of the neurons default to ReLU (change to Softplus if not using laptop). Might need some more tweaking yet.